move to Enterprise tier
As per https://github.com/FlowFuse/flowfuse/issues/2997#issuecomment-1857711941
