+++ forge/routes/api/application.js
@@ -1,5 +1,3 @@
-const applicationDeviceGroup = require('./applicationDeviceGroup.js')
-
 module.exports = async function (app) {
     app.addHook('preHandler', async (request, reply) => {
         const applicationId = request.params.applicationId
@@ -27,8 +25,6 @@ module.exports = async function (app) {
             return reply.code(500).send({ code: 'unexpected_error', error: err.toString() })
         }
     })
-
-    app.register(applicationDeviceGroup, { prefix: '/:applicationId/device-groups' })
 
     /**
      * Create an application
